Created basic app
Also created the fetchWeather hook.
